Trying to make works the copy fuctions, the rest works

+++ ServerPi/plugins/button/button.php
@@ -16,8 +16,11 @@
 			}
 		}
 	?>
-	<a href="ServerPi/plugins/button/reboot.php">
+	<a href="ServerPi/plugins/button/reboot.php" onclick="return(confirm('Are you sure to reboot the system?'));">
 		<div class="reboot"><p>Reboot</p></div>
 	</a>
-	<div class="temp"><p><?php echo shell_exec('awk \'{printf("%.1f°C",$1/1e3)}\' /sys/class/thermal/thermal_zone0/temp'); ?></p></div>
+	<div class="temp" style="background-color: <?php $temp = shell_exec('awk \'{printf("%.1f",$1/1e3)}\' /sys/class/thermal/thermal_zone0/temp'); if($temp > 60) { echo '#f24f5a'; } elseif ($temp > 50) { echo '#e3a42c';} else { echo '#29d251'; } {
+	}?>;">
+		<p><?php echo shell_exec('awk \'{printf("%.1f°C",$1/1e3)}\' /sys/class/thermal/thermal_zone0/temp'); ?></p>
+		</div>
 </div>
